refactor: Computer class 분리
