+++ src/computer/computer.js
@@ -1,0 +1,15 @@
+export default class Computer {
+  generateRandomNumbers() {
+    let answer = [];
+    while (answer.length < NUMBER_LENGTH) {
+      this.addRandomNum(answer);
+    }
+    return answer.join("");
+  }
+  addRandomNum(answer) {
+    const randomNum = String(MissionUtils.Random.pickNumberInRange(1, 9));
+    if (answer.length === 0 || !answer.includes(randomNum)) {
+      answer.push(randomNum);
+    }
+  }
+}
